requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bqplot==0.12.18
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colour==0.1.5
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
earthengine-api==0.1.238
entrypoints==0.3
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Cors==3.0.9
folium==0.11.0
future==0.18.2
geeadd==0.5.2
geemap==0.8.0
geocoder==1.38.1
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
httplib2shim==0.0.3
idna==2.10
importlib-metadata==2.0.0
ipyfilechooser==0.4.0
ipykernel==5.3.4
ipyleaflet==0.13.3
ipynb-py-convert==0.4.6
ipytree==0.1.8
ipywidgets==7.5.1
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyter-server==1.0.5
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
logzero==1.5.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mss==6.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGeoj==1.0.0
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyshp==2.1.2
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
ratelim==0.1.6
requests==2.24.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
voila==0.2.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.3.1
zope.interface==5.1.2
